Worldgen readability refactor — 32 commits.

## What changed

**Submodule splits** (four flat files → four directories):
- caves.rs (2352 lines) → caves/ (12 files)
- hydrology.rs (1148 lines) → hydrology/ (8 files)
- region.rs (666 lines) → region/ (6 files)
- heightmap.rs + density_graph.rs → density/ (6 files)

**Extractions from mod.rs** (426 → 206 lines):
- generator_impl.rs — Generator::new/config/build_fine_region/etc.
- fill_chunk_impl.rs — fill_chunk, light_inputs_for_chunk
- trees_impl.rs — add_trees, tree_in_cell_with_regions, stamp_tree
- probe_impl.rs — probe_column, paint_column, sample_stage
- columns_impl.rs — column_data, column_data_with
- surface_fixer.rs — surface post-passes
- pipeline.rs — ChunkRegions, gather_chunk_regions

**New abstractions** (Phase C):
- RiverWidth(f32), ChamberRadius(Vec3), TunnelRadius(f32)
- SystemBoundingBox, DensityComposition
- TerrainRef, CaveCtx (reduce argument counts)
- WalkCtx + WalkState in carver (14-arg walk() → 2)
- CornerLatticeEvaluator trait (static dispatch, CellEvaluator + CarverEvaluator)

**Documentation**:
- Every pub item has /// rustdoc
- Every tuning constant has Smaller→/Larger→ direction-of-effect
- Every math-heavy section has prose naming the algorithm
- src/worldgen/CLAUDE.md + AGENTS.md added
- All stale file-path references in comments corrected

**Invariant**: worldgen fingerprint hash 0x0652_5CEE_A30F_9710 unchanged.
All 3 fingerprint tests + smoke + gen_with_neighbors pass.

Delete saves/default/regions/ before play-testing — persistent chunk
cache does not invalidate on code changes.

+++ tests/cave_surface_breakthrough.rs
@@ -14,7 +14,7 @@ fn surface_breakthrough_places_grass_on_cave_floor() {
     let mut found_floor_grass = false;
     let mut found_ceiling_grass = false;
 
-    let dim = CHUNK_DIM_U as u32;
+    let dim = CHUNK_DIM_U;
 
     // Scan an 8x8 grid of surface-level chunks across chunk Y 0..=3.
     // Cave breakthroughs are probabilistic; a wider scan provides more
